Follow trunk
This moves the master branch of GoLLVM to closely follow trunk.
Specifically, we use a recent revision of trunk which will be
bumped every so often. We use a specific revision in order to
protect ourselves from unexpected API breakage.

I tried to make it as easy as possible to stay updated. A script,
update_llvm.sh, takes care of fetching the correct revision of LLVM and
building it, but you can also choose to use your own build of LLVM if
you like (with the 'byollvm' tag).

What I had in mind was that this could be the master branch of GoLLVM,
and that we could fork branches from this to correspond with LLVM releases,
where the script would be changed to fetch the relevant tagged release of
LLVM from its repository.